Removing old campaign_info not used. Updating campaign_info mat view refresh from Rogue.
