+++ data/sql/derived-tables/campaign_info.sql
@@ -1,4 +1,5 @@
-CREATE TEMPORARY TABLE IF NOT EXISTS campaign_info_all AS ( 
+DROP MATERIALIZED VIEW IF EXISTS rogue.campaign_info_all;
+CREATE MATERIALIZED VIEW rogue.campaign_info_all AS ( 
     SELECT c.field_campaigns_target_id as campaign_node_id,
            n2.title as campaign_node_id_title,
            c.entity_id as campaign_run_id,
@@ -40,30 +41,30 @@ CREATE TEMPORARY TABLE IF NOT EXISTS campaign_info_all AS (
     GROUP BY 1,2,3,4,5,6,7,8,9,10,11 
     ORDER BY c.field_campaigns_target_id, fdfrd.field_run_date_value);
     
-DROP TABLE IF EXISTS campaign_info;
-CREATE TABLE IF NOT EXISTS campaign_info AS (
+DROP MATERIALIZED VIEW IF EXISTS public.campaign_info;
+CREATE MATERIALIZED VIEW public.campaign_info AS (
 	SELECT 
 		c.id AS campaign_id,
 		c.internal_title AS campaign_name,
 		i.*
-	FROM campaign_info_all i
+	FROM rogue.campaign_info_all i
 	LEFT JOIN rogue_prod.campaigns c ON i.campaign_run_id = c.campaign_run_id
 	WHERE i.campaign_language = 'en'
 );
-GRANT SELECT ON campaign_info TO dsanalyst;
-GRANT SELECT ON campaign_info TO looker;
+GRANT SELECT ON public.campaign_info TO dsanalyst;
+GRANT SELECT ON public.campaign_info TO looker;
 
-DROP TABLE IF EXISTS campaign_info_international;
-CREATE TABLE IF NOT EXISTS campaign_info_international AS (
+DROP MATERIALIZED VIEW IF EXISTS public.campaign_info_international;
+CREATE MATERIALIZED view public.campaign_info_international AS (
 	SELECT 
 		c.id AS campaign_id,
 		c.internal_title AS campaign_name,
 		i.*
-	FROM campaign_info_all i
+	FROM rogue.campaign_info_all i
 	LEFT JOIN rogue_prod.campaigns c ON i.campaign_run_id = c.campaign_run_id
 	WHERE campaign_language IS DISTINCT FROM 'en'
 );
-GRANT SELECT ON campaign_info_international TO dsanalyst;
-GRANT SELECT ON campaign_info_international TO looker;
+GRANT SELECT ON public.campaign_info_international TO dsanalyst;
+GRANT SELECT ON public.campaign_info_international TO looker;
 
 CREATE UNIQUE INDEX ON public.campaign_info (campaign_run_id, campaign_id);
